docs: swap primary structured outputs example
